Fix a broken unit test
This time using the correct value of sample_error_log.

This test isn't passing properly. Looking at the git log,
it seems like it was a typo/mistake when open-sourcing
this project which has gone undiscovered until now.

This is a 1-word change, typing down this long a commit
message feels truly overkill.